requirements.txt
discord.py>=2.0.0
matplotlib>=3.5.0
pytz>=2021.3
python-dotenv>=0.19.0
aiohttp>=3.8.0 